Code Cleanup & Refactoring
