+++ app/app.go
@@ -595,8 +595,7 @@ func (app *App) RegisterAPIRoutes(apiSvr *api.Server, _ config.APIConfig) {
 	// register app's OpenAPI routes.
 	apiSvr.Router.Handle("/static/openapi.yml", http.FileServer(http.FS(docs.Docs)))
 	apiSvr.Router.HandleFunc("/", openapiconsole.Handler(Name, "/static/openapi.yml"))
-	apiSvr.Router.HandleFunc("/webhost", webhost.Handler())
-	apiSvr.Router.HandleFunc("/fs/{contract}/{name}/{path:.*}", webhost.Handler2())
+	apiSvr.Router.HandleFunc("/webhost/{contract}/{name}/{path:.*}", webhost.Handler())
 }
 
 // RegisterTxService implements the Application.RegisterTxService method.
